Change time from uint32_t to unsigned int for large positive values)

+++ tools/sdk/ksdk1.1.0/platform/osa/src/fsl_os_abstraction_free_rtos.c
@@ -708,7 +708,7 @@ void OSA_TimeDelay(uint32_t delay)
  * Description   : This function gets current time in milliseconds.
  *
  *END**************************************************************************/
-uint32_t OSA_TimeGetMsec(void)
+uint64_t OSA_TimeGetMsec(void)
 {
     portTickType ticks;
